+++ src/main/java/com/natoboram/switcheroo/ItemStackUtil.java
@@ -10,6 +10,7 @@ import net.fabricmc.api.Environment;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.enchantment.EnchantmentHelper;
+import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.EntityGroup;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
@@ -51,38 +52,56 @@ public class ItemStackUtil {
 				entityGroup);
 	}
 
-	public static void keepMostDps(final ArrayList<ItemStack> weapons, final EntityGroup entityGroup,
+	public static boolean keepMostDps(final ArrayList<ItemStack> weapons, final EntityGroup entityGroup,
 			@Nullable final Double maxDps) {
 		final double max = maxDps == null ? getMaxDps(weapons, entityGroup) : maxDps.doubleValue();
-		weapons.removeIf(stack -> max > ItemStackUtil.getDps(stack, entityGroup));
+		return weapons.removeIf(stack -> max > ItemStackUtil.getDps(stack, entityGroup));
 	}
 
-	public static void keepFastestTools(final ArrayList<ItemStack> tools, final BlockState blockState) {
-		final float max = tools.stream().max(Comparator.comparing(item -> item.getMiningSpeedMultiplier(blockState)))
-				.get().getMiningSpeedMultiplier(blockState);
-		tools.removeIf(item -> max > item.getMiningSpeedMultiplier(blockState));
+	public static boolean keepFastestTools(final ArrayList<ItemStack> tools, final BlockState blockState) {
+		final double max = getMiningSpeedMultiplier(
+				tools.stream().max(Comparator.comparing(item -> getMiningSpeedMultiplier(item, blockState))).get(),
+				blockState);
+
+		return tools.removeIf(item -> max > getMiningSpeedMultiplier(item, blockState));
 	}
 
-	public static void keepSlowestTools(final ArrayList<ItemStack> tools, final BlockState blockState) {
-		final float min = tools.stream().min(Comparator.comparing(item -> item.getMiningSpeedMultiplier(blockState)))
-				.get().getMiningSpeedMultiplier(blockState);
-		tools.removeIf(item -> min < item.getMiningSpeedMultiplier(blockState));
+	public static boolean keepSlowestTools(final ArrayList<ItemStack> tools, final BlockState blockState) {
+		final double min = getMiningSpeedMultiplier(
+				tools.stream().min(Comparator.comparing(item -> getMiningSpeedMultiplier(item, blockState))).get(),
+				blockState);
+
+		return tools.removeIf(item -> min < getMiningSpeedMultiplier(item, blockState));
 	}
 
-	/** Removes enchanted items that have only one durability left. */
-	public static void removeDamagedEnchantedItems(final ArrayList<ItemStack> items) {
-		items.removeIf(item -> {
-			return !item.getEnchantments().isEmpty() && item.getMaxDamage() - item.getDamage() <= 1;
-		});
+	/**
+	 * Wrapper for {@link ItemStack#getMiningSpeedMultiplier} that takes into
+	 * account {@link Enchantments.EFFICIENCY} levels.
+	 *
+	 * @see <a href= "https://minecraft.fandom.com/wiki/Efficiency">Efficiency</a>
+	 */
+	public static double getMiningSpeedMultiplier(final ItemStack tool, final BlockState blockState) {
+		final float multiplier = tool.getMiningSpeedMultiplier(blockState);
+		if (tool.isSuitableFor(blockState)) {
+			final int level = EnchantmentHelper.getLevel(Enchantments.EFFICIENCY, tool);
+			return multiplier + (level == 0 ? 0 : 1 + Math.pow(level, 2));
+		}
+		return multiplier;
 	}
 
-	public static void keepMostDamagedItems(final ArrayList<ItemStack> items) {
+	/** Removes enchanted items that have only 5 durability left. */
+	public static boolean removeDamagedEnchantedItems(final ArrayList<ItemStack> items, final SwitcherooConfig config) {
+		return items.removeIf(item -> !item.getEnchantments().isEmpty()
+				&& item.getMaxDamage() - item.getDamage() <= config.minDurability);
+	}
+
+	public static boolean keepMostDamagedItems(final ArrayList<ItemStack> items) {
 		final float max = items.stream().max(Comparator.comparing(item -> item.getDamage())).get().getDamage();
-		items.removeIf(item -> max > item.getDamage());
+		return items.removeIf(item -> max > item.getDamage());
 	}
 
-	public static void keepLowestStacks(final ArrayList<ItemStack> items) {
+	public static boolean keepLowestStacks(final ArrayList<ItemStack> items) {
 		final float min = items.stream().min(Comparator.comparing(item -> item.getCount())).get().getCount();
-		items.removeIf(item -> min < item.getCount());
+		return items.removeIf(item -> min < item.getCount());
 	}
 }